# Support comparison functions in `array.sorted`