--- a/Cargo.toml
+++ b/Cargo.toml
@@ -58,6 +58,7 @@ env_proxy = "0.4"
 flate2 = "1"
 fontdb = { version = "0.21", default-features = false }
 fs_extra = "1.3"
+glidesort = "0.1.2"
 hayagriva = "0.8"
 heck = "0.5"
 hypher = "0.1.4"