Closes #5365.

This PR is based on a rebased version of #5371 by @bluebear94. I used `glidesort` instead of the standard library sorting algorithm to prevent panics, which was the blocker for merging #5371.

Original post below.

<hr>

This PR is based on a rebased version of #5371 by @bluebear94. As suggested in https://github.com/typst/typst/pull/5371#issuecomment-2539903091, I implemented a custom sorting function that does not panic, which was the blocker for merging #5371.

## Sorting algorithm

The sorting algorithm I implemented is an in-place stable insertion sort. In particular, it does not allocate anything. The drawback is that it is considerably slower than Rust's default sorting algorithm. I initially tried to implement an in-place stable merge sort,[^1] but was unable to reach a point where I was confident about my implementation.

### Performance

I did a very basic benchmark that involved executing the following function:

```typ
#let test(len) = {
  let k = 0b01010101
  for m in range(256) {
    range(len).map(i => i.bit-xor(k).bit-and(m)).sorted()
  }
}
```

The total execution times of the for loop on my machine, as determined by `--timings`, are shown in the table below.

| `len` | 0.12.0 | This PR |
|-------|--------|---------|
| 1000  | 2.11 s | 3.10 s  |
| 10000 | 15.0 s | 114 s   |

The difference seems quite significant for big arrays. The three solutions I can think of are:

1. Deciding not to care (at the cost of performance).
2. Only using the homemade sorting algorithm when a comparison function is provided (at the cost of performance when a comparison function is provided).
3. Using a better sorting algorithm, such as an in-place stable merge sort (which may be harder to implement in a way that we can be confident is correct).
4. Relying on a crate such as [glidesort](https://crates.io/crates/glidesort) (at the cost of adding a new dependency).

I tend to think option 2. is better.

[^1]: https://thomas.baudel.name/Visualisation/VisuTri/inplacestablesort.html