@xkevio 2024-12-25
Could also use the [`glidesort`](https://github.com/orlp/glidesort/tree/master) instead of re-implementing a way worse sort to keep performance in all cases, mayhaps. This is what driftsort, the new Rust sorting algorithm, is based on but without the panicking total order check.