@MDLC01 2024-12-26
Yes we can also do that, but this requires adding a new dependency.