@xkevio 2025-01-06
> > [...] but without the panicking total order check.
> 
> Is that stated somewhere? I couldn't find anything.

Not explicitly, but you can search for `panic!` or `total order violation` and you will find nothing.