@laurmaedje 2025-01-09
I think I would prefer glidesort over an O(n^2) sort, even if it's an additional dependency.