@laurmaedje 2025-01-10
I think `key` and `compare` wouldn't even need to be mutually exclusive. If there's a key, we could just compare the keys with the comparison function. What do you think?