@laurmaedje 2025-01-10
From a quick look, looks like that's what the code actually does. Then, just the docs need an update.