@MDLC01 2025-01-12
I updated the documentation. There was already a test for the behavior.