@MDLC01 2025-01-23
Quoting the original post from #5371:

> I used `compare` for the argument name instead of `by` because `by` could plausibly refer to either the key or the comparison function.

Of course, I can change it to `by` if you prefer.