@laurmaedje 2025-01-23
I don't know, I like how `by` reads and never found it confusing in Rust.