@MDLC01 2025-01-23
Should we allow `by` to return a float as well?