@laurmaedje 2025-01-23
hmm, when sorting floats it might be useful, but the details around +/- 0 are a bit tricky. if we do, we should first investigate what other languages do.