@MDLC01 2025-01-23
In JavaScript, `-0.0 === 0.0` (and `-0.0` is the valid syntax for creating negative zero). I'm not aware of any other language that allows returning floats from comparison functions (notably, Lua expects a boolean, and Python doesn't even have this feature).