@laurmaedje 2025-01-23
The Lua case is interesting. To make a stable sort possible with just two return values I guess they need to ensure that the the first element passed to the comparison function is always the leftmost one. But if that is a guarantee that one can give, a boolean seems quite straightforward.