@MDLC01 2025-01-23
> But if that is a guarantee that one can give, a boolean seems quite straightforward.

I don't think we can safely assume that if we stick with `glidesort`. At the same time, I agree that returning a boolean seems more user-friendly.