@laurmaedje 2025-01-31
I think booleans would be _much_ more convenient than -1, 0, 1, so I'd prefer to investigate a bit more whether that's somehow possible before proceeding.