@bluebear94 2025-01-31
We can always convert from booleans to orderings using `(b < a) - (a < b)`, but that requires two calls to the comparison function, so I don’t think that’s a good idea.