@MDLC01 2025-02-28
I found a way to rely on `glidesort` while using a boolean comparison function, without calling it twice for each comparison.

The idea is that when a sorting algorithm calls a comparison function, there are only two possible outcomes: the elements are swapped, or they are not. This means out of the three outcomes (`<`, `=` or `>`), two (in practice, either `<` and `=`, or `=` and `>`) yield the same result. Essentially, for each comparison, the sorting algorithm only cares about either `<=` or `>=`. If we know which one for each comparison, we can just decide to pass the two arguments to `by` in whatever order gives us the desired result (as `by` computes `<=`, and therefore swapping its arguments computes `>=`). This is explained more formally in the code.

The drawback is that we need to make an extra allocation. This could be prevented when `by` is not passed, but would likely require duplicating some code. I think it's fine that way (`glidesort` allocates either way).

Additionally, I believe this method effectively makes any sorting algorithm stable. So maybe we can find even better sorting algorithms if we also include unstable sorting algorithms.