@MDLC01 2025-03-31
I initially did not do that because I feared it would have made the code even more complex. In the end, it's fine, but I'm not sure the performance impact is very noticeable. I may try to do a benchmark later if I have the time, but don't count on it.